backend/investment-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.30
aiomysql==0.2.0
pydantic-settings==2.5.0
email-validator==2.2.0
pytest==8.3.5
pytest-asyncio==0.25.3
